+++ website/app/docs/Cache/_meta.js
@@ -1,0 +1,4 @@
+export default {
+  LruCache: 'LruCache',
+  TtlCache: 'TtlCache',
+};